unify on firebase api key
