+++ .env.example
@@ -1,0 +1,8 @@
+# Server
+FLYWHEEL_FIREBASE_PROJECT_ID="firebase-project-id"
+FLYWHEEL_FIREBASE_API_KEY="firebase-api-key"
+
+# Remote Client
+FLYWHEEL_SERVER_HOSTNAME="flywheel.example.com"
+FLYWHEEL_SERVER_TCP_PORT="8888"
+FLYWHEEL_SERVER_UDP_PORT="8889"
